### 2024-11-22 6f2b904
ENH: dvs_nmost and dvs_max return same type as input
