+++ pyproject.toml
@@ -50,7 +50,7 @@ test = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.8.4",
+    "ruff==0.8.6",
 ]
 dev = [
     "cogapp",
@@ -60,7 +60,7 @@ dev = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.8.4",
+    "ruff==0.8.6",
 ]
 doc  = ["click",
         "ipykernel",
