### 2025-01-10 #7503 @Kevin-Mattheus-Moerman
@GavinHuttley Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Apologies for the delay encountered so far. We have a backlog of submissions to process so we are a bit slower than usual to assign a handling editor. I hope to invite one today.

Before we proceed, please can you have a look at the following points:
