### 2025-01-12 3bdfeb1
DOC: updated bib to include doi
