+++ paper/paper.bib
@@ -1,13 +1,99 @@
 
-@article{parks.2018.natbiotechnol,
-	author = {Parks, Donovan H. and Chuvochina, Maria and Waite, David W. and Rinke, Christian and Skarshewski, Adam and Chaumeil, Pierre-Alain and Hugenholtz, Philip},
-	journal = {Nature Biotechnology},
-	month = nov,
-	number = {10},
-	pages = {996--1004},
-	title = {A standardized bacterial taxonomy based on genome phylogeny substantially revises the tree of life},
-	volume = {36},
-	year = {2018}}
+@article{balaban.2019.plosone,
+	author = {Balaban, Metin and Moshiri, Niema and Mai, Uyen and Jia, Xingfan and Mirarab, Siavash},
+	doi = {10.1371/journal.pone.0221068},
+	editor = {Bozdag, Serdar},
+	issn = {1932-6203},
+	journal = {PLOS ONE},
+	language = {en},
+	month = aug,
+	number = {8},
+	pages = {e0221068},
+	shorttitle = {{TreeCluster}},
+	title = {{TreeCluster}: {Clustering} biological sequences using phylogenetic trees},
+	url = {https://dx.plos.org/10.1371/journal.pone.0221068},
+	urldate = {2024-08-14},
+	volume = {14},
+	year = {2019}}
+
+@article{widmann.2006.molcellproteomics,
+	author = {Widmann, Jeremy and Hamady, Micah and Knight, Rob},
+	doi = {10.1074/mcp.T600022-MCP200},
+	journal = {Mol Cell Proteomics},
+	language = {eng},
+	number = {8},
+	pages = {1520--1532},
+	pmid = {16769708},
+	title = {{DivergentSet}, a tool for picking non-redundant sequences from large sequence collections.},
+	volume = {5},
+	year = {2006}}
+
+@article{schneider.1990.nucleicacidsres,
+	author = {Schneider, T D and Stephens, R M},
+	doi = {10.1093/NAR/18.20.6097},
+	issn = {0305-1048},
+	journal = {Nucleic acids research},
+	month = oct,
+	number = {20},
+	pages = {6097--100},
+	pmid = {2172928},
+	title = {Sequence logos: a new way to display consensus sequences.},
+	url = {http://www.ncbi.nlm.nih.gov/pubmed/2172928},
+	urldate = {2016-11-08},
+	volume = {18},
+	year = {1990}}
+
+@article{lin.1991.ieeetrans.inf.theory,
+	author = {Lin, J.},
+	doi = {10.1109/18.61115},
+	issn = {1557-9654},
+	journal = {IEEE Transactions on Information Theory},
+	month = jan,
+	number = {1},
+	pages = {145--151},
+	title = {Divergence measures based on the {Shannon} entropy},
+	url = {https://ieeexplore.ieee.org/document/61115},
+	urldate = {2024-08-21},
+	volume = {37},
+	year = {1991}}
+
+@article{lake.1994.procnatlacadsciua,
+	author = {Lake, J A},
+	journal = {Proc Natl Acad Sci U S A},
+	language = {eng},
+	number = {4},
+	pages = {1455--1459},
+	pmid = {8108430},
+	title = {Reconstructing evolutionary trees from {DNA} and protein sequences: paralinear distances.},
+	url = {https://pubmed.ncbi.nlm.nih.gov/8108430},
+	volume = {91},
+	year = {1994}}
+
+@article{harrison.2024.nucleicacidsresearch,
[…]
