+++ pyproject.toml
@@ -50,7 +50,7 @@ test = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.8.6",
+    "ruff==0.9.1",
 ]
 dev = [
     "cogapp",
@@ -60,7 +60,7 @@ dev = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.8.6",
+    "ruff==0.9.1",
 ]
 doc  = ["click",
         "ipykernel",
