### 2025-01-19 932336f
STY: pre-commit linting with ruff
