### 2025-01-23 #7503 @Kevin-Mattheus-Moerman
@fboehm can you handle this one?
