### 2025-01-26 #7503 @fboehm
@jvkersch @lskatz @marekborowiec @xin-huang @rraadd88 - My name is Fred Boehm, and I'm an associate editor for JOSS. Would you be willing to review this submission for JOSS? Please let me know if you have any questions. Thank you!
