### 2025-01-27 25d4c5a
MAINT: correct __array__ signature
