### 2025-01-27 906b612
STY: yupe hints on more funcs / methods
