+++ src/diverse_seq/record.py
@@ -4,9 +4,9 @@ import dataclasses
 import functools
 from collections.abc import Iterator
 from math import fabs, isclose
-from typing import Union
 
 import numba
+import typing_extensions
 from attrs import asdict, define, field, validators
 from cogent3 import get_moltype
 from cogent3.app import composable
@@ -30,7 +30,7 @@ from numpy import divmod as np_divmod
 
 from diverse_seq import util as dvs_utils
 
-NumType = Union[float, int]
+NumType = float | int
 PosDictType = dict[int, NumType]
 
 
@@ -38,8 +38,9 @@ PosDictType = dict[int, NumType]
 def _gettype(name) -> type:
     try:
         return name.type
-    except AttributeError:
-        raise TypeError(f"type {type(name)} not supported")
+    except AttributeError as e:
+        msg = f"type {type(name)} not supported"
+        raise TypeError(msg) from e
 
 
 @_gettype.register
@@ -64,7 +65,7 @@ class lazy_kmers:
     num_states: int = dataclasses.field(init=False)
     moltype: dataclasses.InitVar[str] = dataclasses.field(default="dna")
 
-    def __post_init__(self, moltype: str):
+    def __post_init__(self, moltype: str) -> None:
         self.num_states = len(_get_canonical_states(moltype))
 
     def __array__(
@@ -125,7 +126,7 @@ class vector:
         dtype: type = float,
         source: str = None,
         name: str = None,
-    ):
+    ) -> None:
         """
         Parameters
         ----------
@@ -145,7 +146,7 @@ class vector:
         self.source = source
         self.name = name
 
-    def __setitem__(self, index: int, value: NumType):
+    def __setitem__(self, index: int, value: NumType) -> None:
         self.data[index] = value
 
     def __getitem__(self, index: int) -> NumType:
@@ -157,15 +158,15 @@ class vector:
     def __iter__(self) -> Iterator[NumType]:
         yield from self.data
 
-    def __getstate__(self):
+    def __getstate__(self) -> dict:
         return asdict(self)
 
-    def __setstate__(self, data):
+    def __setstate__(self, data: dict) -> typing_extensions.Self:
         for k, v in data.items():
             setattr(self, k, v)
         return self
 
-    def __sub__(self, other):
+    def __sub__(self, other: typing_extensions.Self) -> typing_extensions.Self:
         data = self.data - other
         return self.__class__(
             data=data,
@@ -173,11 +174,11 @@ class vector:
             dtype=self.dtype,
         )
 
-    def __isub__(self, other):
+    def __isub__(self, other: typing_extensions.Self) -> typing_extensions.Self:
         self.data -= other
         return self
 
-    def __add__(self, other):
+    def __add__(self, other: typing_extensions.Self) -> typing_extensions.Self:
         # we are creating a new instance
         data = self.data + other
         return self.__class__(
@@ -186,31 +187,31 @@ class vector:
             dtype=self.dtype,
         )
 
-    def __iadd__(self, other):
+    def __iadd__(self, other: typing_extensions.Self) -> typing_extensions.Self:
         self.data += other
         return self
 
-    def __truediv__(self, other):
+    def __truediv__(self, other: typing_extensions.Self) -> typing_extensions.Self:
         # we are creating a new instance
         with errstate(divide="ignore", invalid="ignore"):
[…]
