### 2025-01-27 a6b1b03
ENH: enable support for python 3.13
