### 2025-01-27 bf829be
DEV: use uv backend with nox
