### 2025-01-27 1bc5e43
DEV: updates to tooling settings
