+++ .github/workflows/release.yml
@@ -19,10 +19,16 @@ jobs:
         with:
             python-version: "${{ matrix.python-version }}"
 
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+        with:
+          enable-cache: true
+          cache-dependency-glob: "pyproject.toml"
+
       - name: "Installs for ${{ matrix.python-version }}"
         run: |
-          pip install --upgrade pip
-          pip install nox uv
+          uv venv venv -p python${{ matrix.python-version }}
+          uv tool install -p venv nox
 
       - name: "Run nox for Python ${{ matrix.python-version }}"
         run: "nox -db uv -s test-${{ matrix.python-version }}"
