+++ pyproject.toml
@@ -52,7 +52,7 @@ test = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.9.3",
+    "ruff==0.9.4",
 ]
 dev = [
     "cogapp",
@@ -62,7 +62,7 @@ dev = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.9.3",
+    "ruff==0.9.4",
 ]
 doc  = ["click",
         "ipykernel",
