### 2025-02-03 6e3e462
Bump astral-sh/setup-uv from 3 to 5
Bumps [astral-sh/setup-uv](https://github.com/astral-sh/setup-uv) from 3 to 5.
- [Release notes](https://github.com/astral-sh/setup-uv/releases)
- [Commits](https://github.com/astral-sh/setup-uv/compare/v3...v5)

---
updated-dependencies:
- dependency-name: astral-sh/setup-uv
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
