+++ .github/workflows/ci.yml
@@ -25,7 +25,7 @@ jobs:
             python-version: "${{ matrix.python-version }}"
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v3
+        uses: astral-sh/setup-uv@v5
         with:
           enable-cache: true
           cache-dependency-glob: "pyproject.toml"
