### 2025-02-06 #7503 @fboehm
@iimog @gavinmdouglas @tamarahadz @sdhutchins - would you be willing to review this submission for JOSS?
