### 2025-02-10 #7765 @editorialbot
[REVIEW]: `diverse-seq`: an application for alignment-free selecting and clustering biological sequences
**Submitting author:** @GavinHuttley (<a href="http://orcid.org/0000-0001-7224-2074">Gavin Huttley</a>)
**Repository:** https://github.com/HuttleyLab/DiverseSeq
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 2024.11.8a3
