### 2025-02-24 HuttleyLab/DiverseSeq#110 @xin-huang
[JOSS REVIEW] Suggestions for Documentation
This is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/7765)

Currently, the README in the GitHub repository does not include installation instructions. The only mention of installation is in the paper, which claims that `diverse-seq` can be installed from the Python Package Index.

However, when I tried:

```
pip install diverse-seq
```

or

```
pip install diverse_seq
```

Both commands resulted in the following error:

```
ERROR: Could not find a version that satisfies the requirement diverse-seq (from versions: none)
ERROR: No matching distribution found for diverse-seq
```

or

```
ERROR: Could not find a version that satisfies the requirement diverse_seq (from versions: none)
ERROR: No matching distribution found for diverse_seq
```

The authors should provide clear installation instructions as required by the review checklist and explicitly list all dependencies of the program.

Additionally, the documentation lacks usage examples. It only provides command-line options without demonstrating how to use them, making it difficult for users to understand the software’s functionality. The authors should include concrete examples to assist users.

Moreover, there is no API documentation. The authors should add proper API documentation as required.

Finally, there are no clear guidelines for third parties who wish to “Report issues or problems with the software” or “Seek support.” The authors should provide clear instructions for these aspects to ensure better community engagement.
