### 2025-02-24 HuttleyLab/DiverseSeq#110 @xin-huang
I also noticed that the authors use both `diverse-seq` and `diverse_seq` in the README.

https://github.com/HuttleyLab/DiverseSeq/blob/43252a63ef7561aea4a2d70ddcfe0b014fcd5f09/README.md?plain=1#L7

https://github.com/HuttleyLab/DiverseSeq/blob/43252a63ef7561aea4a2d70ddcfe0b014fcd5f09/README.md?plain=1#L9

https://github.com/HuttleyLab/DiverseSeq/blob/43252a63ef7561aea4a2d70ddcfe0b014fcd5f09/README.md?plain=1#L11

I suggest maintaining consistency in naming to avoid confusion.
