### 2025-02-24 HuttleyLab/DiverseSeq#110 @GavinHuttley
Hi @xin-huang ,
Thanks for pointing out the inconsistent naming in the README. Fixed in the main branch and will merge into JOSS branch soon. We will also address the additional documentation request soon.

In terms of the installation issue, I confirm that both `pip install diverse-seq` and `pip install diverse_seq` work in a conda virtual environment with python 3.12.

The minimum supported python version is python3.10. Can you provide more details on the environment, including OS and version, that failed for you. Thanks!
