### 2025-02-24 HuttleyLab/DiverseSeq#110 @GavinHuttley
Thanks. I think the real issue is the installation error from `pip` is not informative. If the installation was done using [uv pip](https://docs.astral.sh/uv/#the-pip-interface) then the error message is informative, e.g.
```
uv pip install diverse-seq
Using Python 3.9.21 environment at: /Users/gavin/miniconda3/envs/delme
  × No solution found when resolving dependencies:
  ╰─▶ Because the current Python version (3.9.21) does not satisfy Python>=3.10,<3.14 and diverse-seq<=2024.12.26a1 depends on Python>=3.10,<3.13, we can conclude that
      diverse-seq<=2024.12.26a1 cannot be used.
      And because only the following versions of diverse-seq are available:
          diverse-seq==2024.8.26a5
          diverse-seq==2024.8.26a6
          diverse-seq==2024.9.2a1
          diverse-seq==2024.11.8a1
          diverse-seq==2024.11.8a2
          diverse-seq==2024.11.8a3
          diverse-seq==2024.11.22a1
          diverse-seq==2024.12.26a1
          diverse-seq==2024.12.26a2
      we can conclude that diverse-seq<2024.12.26a2 cannot be used....
```
I will recommend users try that if they run into installation issues.

- [ ] Perhaps I can address the minimum version issue by adding a badge that displays the supported python versions
- [ ] happy to link to the `pyproject.toml` to display the dependencies
