### 2025-02-28 HuttleyLab/DiverseSeq#116 @xin-huang
[JOSS REVIEW] Suggestions for Software Paper
This is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/7765)

**1. State of the Field Section**
The manuscript is missing a "State of the Field" section, which should describe how this software compares to other commonly used packages. However, this comparison is partially covered in the "Statement of Need" section.

https://github.com/HuttleyLab/DiverseSeq/blob/17a5000b185bf3437a7ea56e01ecbe074ccfb2a7/paper/paper.md?plain=1#L49-L51

Additionally, some sentences in the "Summary" section could be moved to "Statement of Need" for better organization.

https://github.com/HuttleyLab/DiverseSeq/blob/17a5000b185bf3437a7ea56e01ecbe074ccfb2a7/paper/paper.md?plain=1#L37

To align with the review checklist, I suggest restructuring these sections.

**2. Simulated Data and Reproducibility**
The authors state that they used simulated data to evaluate the performance of `dvs max`.

https://github.com/HuttleyLab/DiverseSeq/blob/17a5000b185bf3437a7ea56e01ecbe074ccfb2a7/paper/paper.md?plain=1#L106

but I could not find these data in the submission. Additionally, the manuscript does not describe how the simulated data were generated. To ensure reproducibility, the authors should provide details on the data generation process and make the simulated data available.

**3. Analysis Scripts and Workflow Management**
Since the manuscript presents multiple analyses, I suggest including the scripts used for these analyses. Ideally, using a workflow manager such as Nextflow or Snakemake would improve reproducibility and serve as practical examples for users to understand the package.

**4. Figure Placement**
The figures are currently placed at the end of the manuscript. To improve readability and facilitate easier reference, I recommend moving them into the main text, positioning each figure near its first mention.

**5. Zenodo Link Formatting**
I suggest converting the Zenodo reference into a clickable hyperlink to enhance accessibility and convenience for readers.

https://github.com/HuttleyLab/DiverseSeq/blob/17a5000b185bf3437a7ea56e01ecbe074ccfb2a7/paper/paper.md?plain=1#L138
