### 2025-03-01 #7765 @xin-huang
In this manuscript, the authors introduce `diverse-seq`, a package for selecting representative sequences from a collection without requiring alignment. This tool has the potential to benefit the community by facilitating phylogenetic inference from large-scale datasets. However, the current manuscript does not fully meet the requirements outlined in the checklist. Therefore, I have opened two issues, one for the documentation (https://github.com/HuttleyLab/DiverseSeq/issues/110) and another for the software paper (https://github.com/HuttleyLab/DiverseSeq/issues/116), to address specific concerns.
