+++ pyproject.toml
@@ -52,7 +52,7 @@ test = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.9.7",
+    "ruff==0.9.9",
 ]
 dev = [
     "cogapp",
@@ -62,7 +62,7 @@ dev = [
     "pytest",
     "pytest-cov",
     "pytest-xdist",
-    "ruff==0.9.7",
+    "ruff==0.9.9",
 ]
 doc  = ["click",
         "ipykernel",
