### 2025-03-04 88d6830
STY: pre-commit linting with ruff
