### 2025-03-12 096fa99
ENH: added sample data file and functions for getting it
[NEW] dvs demo-data will write out a sample alignment file
    for users to try out the Software

[NEW] diverse_seq.load_sample_data() returns a cogent3 SequenceCollection
    for trialling the software within a python script.
