### 2025-03-12 e59348e
ENH: cli now supports for specifying the random number seed
[CHANGED] the shulffing is done on the seqids before calling
    the different functions
