### 2025-03-12 d02db0c
STY: improve errors and warnings
