### 2025-03-12 f350f79
MAINT: accept sourcery-ai suggestion to allow a random seed of 0
