+++ docs/index.md
@@ -1,0 +1,17 @@
+# `diverse-seq` provides alignment-free algorithms to facilitate phylogenetic workflows
+
+`diverse-seq` implements computationally efficient alignment-free algorithms that enable efficient prototyping for phylogenetic workflows. It can accelerate parameter selection searches for sequence alignment and phylogeny estimation by identifying a subset of sequences that are representative of the diversity in a collection.
+
+You can read more about the methods implemented in `diverse-seq` in the preprint [here](https://biorxiv.org/cgi/content/short/2024.11.10.622877v1).
+
+### Installation
+
+`diverse-seq` can be installed from PyPI as follows
+
+```
+pip install diverse-seq
+```
+
+> **NOTE**
+> If you experience any errors during installation, we recommend using [uv pip](https://docs.astral.sh/uv/). This command provides much better error messages than the standard `pip` command. If you cannot resolve the installation problem, please [open an issue](https://github.com/HuttleyLab/DiverseSeq/issues).
+
