+++ .readthedocs.yaml
@@ -9,12 +9,8 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.12"
+    python: "3.13"
   commands:
-    # Install the required dependencies
-    - pip install requests
-    # Run the script to download and extract the pre-built docs
-    - python rtd_get_docs.py
     - echo "Documentation downloaded and extracted"
 
 mkdocs:
