### 2025-03-12 b4edb10
DEV: actually build the docs ... doh!
