### 2025-03-12 537fd57
DEV: add mkdocs to doc group
