### 2025-03-12 1a201d7
DEV: cleanup the doc dependencies
