+++ pyproject.toml
@@ -58,7 +58,6 @@ doc  = ["click",
         "ipykernel",
         "ipython",
         "ipywidgets",
-        "jupyter-sphinx",
         "jupyter_client",
         "jupyterlab",
         "jupytext",
@@ -69,13 +68,8 @@ doc  = ["click",
         "markdown>=3.7",
         "nbconvert>5.4",
         "nbformat",
-        "nbsphinx",
         "numpydoc",
-        "piqtree",
-        "pandas",
         "plotly",
-        "requests",
-        "statsmodels",
         ]
 dev = [
     "cogapp",
