### 2025-03-12 38d0b7f
DOC: links to rtd page and update README, fixes #110
