### 2025-03-13 91ecbdf
DOC: added README to paper nbks directory
