### 2025-03-13 5cb6918
DOC: address reviewer comments in manuscript, related to #116
