### 2025-03-13 HuttleyLab/DiverseSeq#116 @GavinHuttley
Hi @xin-huang,

Thank you so much for your very thorough comments! I have updated the documentation, as you suggested.

**1. Absence of the "State of the field" section**

I note that the [JOSS sample paper](https://joss.readthedocs.io/en/latest/example_paper.html) does not contain that section, but has "Statement of need" as does [a recent paper](https://joss.theoj.org/papers/10.21105/joss.07421). So we will retain that.

- [x] We have modified the Summary section as suggested.

**2. Simulated data and reproducibility**

We don't include the data we used in producing the paper figures because of their size. We do specify that all the scripts and notebooks used to produce figures are part of the GitHub repository, but admittedly that statement should be more explicit. These are under the JOSS branch at https://github.com/HuttleyLab/DiverseSeq/tree/JOSS/paper with scripts in `nbks/`, raw results in `results/` and the preprint figures in `figs/`.

- [x] We now provide a direct link to the GitHub `nbks/` in the manuscript.

- [x] Your point about reproducibility is well made. In addition to being more explicit about where the code and sample data are location, we have added to our command line tools a `--seed` option to set a random seed. Running the commands in verbose mode will display the seed used whether set by the user or not. A statement to this effect is included in [the documentation](https://diverse-seq.readthedocs.io/en/latest/cli/#selecting-the-maximally-diverse-sequences-using-max-command). We have added a new command to write out a sample data file and show, in the new documentation, using that file for the different commands.

- [x] We have included a new [nbks/README.md](https://github.com/HuttleyLab/DiverseSeq/blob/JOSS/paper/nbks/README.md) file which states how to get setup to reproduce the analyses a script to get the data.

**4. Figure Placement**

We have followed the JOSS template which has everything at the end. Like you, we find it more natural to embed tables and figures near first usage. At this point, we leave as is until instructed to revise by the JOSS editors.

**5. Zenodo Link Formatting**

- [x] We have fixed the Zenodo link formatting.
