### 2025-03-20 #7765 @iimog
> [@iimog](https://github.com/iimog) - just wanted to check to see how the review is going. Please let me know if there are any difficulties or any evaluation criteria need clarifications. Thanks again!

I'm on it. I'm currently collecting my points and will open a bunch of issues once I'm done. I'm doing it this way, to avoid reviewing a moving target - is it preferred to open the issues one by one?
