### 2025-03-21 HuttleyLab/DiverseSeq#134 @iimog
[JOSS REVIEW] Software Paper
Hi, I'm Markus, one of the reviewers for JOSS (https://github.com/openjournals/joss-reviews/issues/7765).
In this issue, I summarize my observations regarding the software paper. Overall, I found the manuscript very well-written. So I only have some minor comments:

1. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L49
[...] tools to the [...] → tools for the

2. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L55
The example with `H=2` is only true for `k=1`, e.g. for `k=2`, the sequence with equifrequent nucleotides `ACGTACGT` is not 2.

3. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L95
I don't understand the meaning of this sentence: "This subset, the MinHash sketch, is the *sketch size* smallest $k$-mers when sorted by a hash function."

4. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L106
You simulated 50 sequences, of which one sequence occurred at 1%. Does this mean that it was not present some of the time? Because occurring once would already be a frequency of 2%.

5. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L151
Figure 5 is never referenced from the main text. Also, there is no Figure 1 in the pdf.

6. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L163
Maybe specify "relationship *of nucleotide frequencies*". I  was a bit confused because I thought it is the relationship of the sequences and the sequences themselves are not identical in the first row.

7. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L130
Consider adding a cdot between "s" and "log l" to make it clearer that it is $s \log s \cdot \log l$ and not $s \log (s \log l)$.

8. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L147
Figure 3: "cov" and "stdev" should be explained in the figure caption.

9. https://github.com/HuttleyLab/DiverseSeq/blob/b897499f7591bbd079fdf6c2fdc21a4b1e09c9da/paper/paper.md?plain=1#L149
Figure 4: the color scheme is very similar to Figure 3 but the mapping is different (std is blue and cov red here, and the reverse in the previous figure)
