### 2025-03-21 HuttleyLab/DiverseSeq#135 @iimog
Use of -hp in documentation
The parameter `-hp` is used throughout the documentation: https://diverse-seq.readthedocs.io/en/latest/cli/
I assume that is because, you don't want to include the progress bars in the output of the commands. However, most users probably want the default behavior and see the progress bars in most cases. Is it possible to set this using an environment variable so `mkdocs` does not produce progress bars even without setting this option? Otherwise, the `-hp` flag should be explained in the `prep` command, currently it is only explained in the `nmost` command. It might also make sense to use the long form of options so users know what is being specified.
