### 2025-03-21 HuttleyLab/DiverseSeq#136 @iimog
Add installation instruction via uv
As you already mention `uv` in the installation section you might consider adding instructions to install just the cli via

```bash
uv tool install diverse-seq
```

The advantage is, that you don't need a python environment that also contains the python module afterwards. It's also possible to quickly test the tool without installation using

```bash
uvx --from diverse-seq dvs
```

None of this is necessary but I find it quite useful.
