### 2025-03-21 HuttleyLab/DiverseSeq#137 @iimog
Option -s of dvs prep is confusing
The documentation for -s when running `dvs prep --help` is:
```
  -s, --seqdir PATH        directory containing sequence files  [required]
```

But in the docs https://diverse-seq.readthedocs.io/en/latest/cli/#the-prep-command a single fasta file, not a directory is provided. Moreover, when providing the directory containing `demo.fa` the prep command seems to work and produces a .dvseqs output. But this file can not be used for subsequent commands (e.g. `dvs nmost`) produces this error:

```
$ dvs nmost -s demo.dvseqs -o demo-nmost.tsv -k 6 -n 10 -hp
NotCompleted(type=ERROR, origin=select_nmost, source="Unknown", message="Traceback (most recent call last):
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/cogent3/app/composable.py", line 401,
in _call
    result = self.main(val, *args, **kwargs)
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/diverse_seq/records.py", line 561, in
main
    return most_divergent(records, size=self._max_size, verbose=self._verbose > 0)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/diverse_seq/records.py", line 394, in
most_divergent
    sr = SummedRecords.from_records(records[:size])
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/diverse_seq/records.py", line 158, in
from_records
    _delta_jsd(summed_kfreqs, summed_entropies, records),
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/diverse_seq/records.py", line 100, in
_delta_jsd
    jsd = _jsd(summed_kfreqs, summed_entropies, n - 1)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/miniforge3/lib/python3.12/functools.py", line 909, in wrapper
    return dispatch(args[0].__class__)(*args, **kw)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/tank/home/iimog/.local/share/uv/tools/diverse-seq/lib/python3.12/site-packages/diverse_seq/records.py", line 68, in
_
    entropy = summed_entropy / n
              ~~~~~~~~~~~~~~~^~~
ZeroDivisionError: float division by zero
")
ERROR: invalid data type, 'NotCompleted' not in SummedRecords
```

Related to this, the description "[...] the first is the name of the file the sequence came from [...]" in https://diverse-seq.readthedocs.io/en/latest/cli/#the-nmost-command is confusing as following the demo, all sequences come from the same file (demo.fa).
