### 2025-03-21 HuttleyLab/DiverseSeq#138 @iimog
dvs prep seems to overwrite even without -F
**Describe the bug**
When I run `dvs prep` with an existing output file but without `-F` I get the message
> A file with the same name already exists. Existing data members will be skipped. Use the -F flag if you want to overwrite the existing file.

However, the dvseqs file seems to be completely overwritten.

**To Reproduce**
Steps to reproduce the behavior:
1. Create `demo.fa` via `dvs demo-data`
2. Run `dvs prep -s demo.fa -o demo.dvseqs`
3. Create a smaller file with a single sequence `head -n 20 demo.fa >head.fa`
4. Run `dvs prep -s head.fa -o demo.dvseqs`
5. See the message but verify with `h5ls demo.dvseqs` that it only contains one sequence (FlyingFox with length 1140 bp)
6. Run `dvs prep -s demo.fa -o demo.dvseqs` again
7. It contains all sequences again but also FlyingFox is back at 2789 bp.

**Expected behavior**
Based on the message I expected files that are already in the database to not be changed. So after the initial `dvs prep` the file should not change in the example above.

 - OS: Debian
 - Version: dvs, version 2025.3.13
