### 2025-03-21 HuttleyLab/DiverseSeq#139 @iimog
Last example in python docu requires plotly, kaleido, and nbformat to be installed
When following the Python API docs https://diverse-seq.readthedocs.io/en/latest/apps/#using-dvs_ctree in a fresh environment that only contains `diverse-seq`, plotting the tree in the last example fails:

```python
dnd = result.get_figure()
dnd.show(renderer="svg", height=700, width=400)
```

```
Cell In[18], line 2
      1 dnd = tree.get_figure()
----> 2 dnd.show(renderer="svg", height=700, width=400)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:197, in Drawable.show(self, renderer, **kwargs)
    195 @extend_docstring_from(_show_)
    196 def show(self, renderer=None, **kwargs) -> None:
--> 197     _show_(self, renderer, **kwargs)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:62, in _show_(cls, renderer, **kwargs)
     47 def _show_(cls, renderer=None, **kwargs) -> None:
     48     """display figure
     49
     50     Parameters
   (...)     60         other arguments for plotly.io.show
     61     """
---> 62     from plotly.io import show
     64     if renderer is None and PLOTLY_RENDERER is None:
     65         renderer = "notebook_connected+plotly_mimetype"

ModuleNotFoundError: No module named 'plotly'
```

Installing `plotly` leads to the next error:
```
Cell In[31], line 2
      1 dnd = tree.get_figure()
----> 2 dnd.show(renderer="svg", height=700, width=400)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:197, in Drawable.show(self, renderer, **kwargs)
    195 @extend_docstring_from(_show_)
    196 def show(self, renderer=None, **kwargs) -> None:
--> 197     _show_(self, renderer, **kwargs)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:80, in _show_(cls, renderer, **kwargs)
     78 kwargs["width"] = fig.layout.width = width
     79 kwargs["height"] = fig.layout.height = height
---> 80 show(fig, **kwargs)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/plotly/io/_renderers.py:417, in show(fig, renderer, validate, **kwargs)
    414 fig_dict = validate_coerce_fig_to_dict(fig, validate)
    416 # Mimetype renderers
--> 417 bundle = renderers._build_mime_bundle(fig_dict, renderers_string=renderer, **kwargs)
    418 if bundle:
    419     if not ipython_display:

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/plotly/io/_renderers.py:325, in RenderersConfig._build_mime_bundle(self, fig_dict, renderers_string, **kwargs)
    322             if hasattr(renderer, k):
    323                 setattr(renderer, k, v)
--> 325         bundle.update(renderer.to_mimebundle(fig_dict))
    327 return bundle

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/plotly/io/_base_renderers.py:126, in ImageRenderer.to_mimebundle(self, fig_dict)
    125 def to_mimebundle(self, fig_dict):
--> 126     image_bytes = to_image(
    127         fig_dict,
    128         format=self.format,
    129         width=self.width,
    130         height=self.height,
    131         scale=self.scale,
    132         validate=False,
    133         engine=self.engine,
    134     )
    136     if self.b64_encode:
    137         image_str = base64.b64encode(image_bytes).decode("utf8")

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/plotly/io/_kaleido.py:132, in to_image(fig, format, width, height, scale, validate, engine)
    130     # Raise informative error message if Kaleido is not installed
    131     if scope is None:
--> 132         raise ValueError(
    133             """
    134 Image export using the "kaleido" engine requires the kaleido package,
    135 which can be installed using pip:
    136     $ pip install -U kaleido
    137 """
    138         )
    140     # Validate figure
    141     # ---------------
    142     fig_dict = validate_coerce_fig_to_dict(fig, validate)

ValueError:
Image export using the "kaleido" engine requires the kaleido package,
which can be installed using pip:
    $ pip install -U kaleido
```
Also installing `kaleido` led to the next problem:
```
Cell In[10], line 2
      1 dnd = tree.get_figure()
----> 2 dnd.show(renderer="svg", height=700, width=400)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:197, in Drawable.show(self, renderer, **kwargs)
    195 @extend_docstring_from(_show_)
    196 def show(self, renderer=None, **kwargs) -> None:
--> 197     _show_(self, renderer, **kwargs)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/cogent3/draw/drawable.py:80, in _show_(cls, renderer, **kwargs)
     78 kwargs["width"] = fig.layout.width = width
     79 kwargs["height"] = fig.layout.height = height
---> 80 show(fig, **kwargs)

File ~/miniforge3/envs/review-joss-diverse-seq/lib/python3.13/site-packages/plotly/io/_renderers.py:425, in show(fig, renderer, validate, **kwargs)
    420     raise ValueError(
    421         "Mime type rendering requires ipython but it is not installed"
    422     )
    424 if not nbformat or Version(nbformat.__version__) < Version("4.2.0"):
--> 425     raise ValueError(
    426         "Mime type rendering requires nbformat>=4.2.0 but it is not installed"
    427     )
    429 display_jupyter_version_warnings()
    431 ipython_display.display(bundle, raw=True)

ValueError: Mime type rendering requires nbformat>=4.2.0 but it is not installed
```

Also installing `nbformat` solves the problem. This should be mentioned in the documentation.

In case that's important I created the environment as follows and I'm using notebooks in vscode (which automatically installs the ipykernel package into the environment):
```
mamba create -n review-joss-diverse-seq python
mamba activate review-joss-diverse-seq
uv pip install diverse-seq
```
