### 2025-03-21 #7765 @iimog
I'm done with my review.
The authors present `diverse-seq` a python tool (available as module and cli) which can perform tasks in molecular evolutionary biology: selecting diverse/representative sequences, and estimating a phylogenetic tree. The methods are well-described, and the tool is mature, well-tested, and well-documented. I opened an issue with some minor comments regarding the paper:
- https://github.com/HuttleyLab/DiverseSeq/issues/134

Furthermore, I opened some issues regarding documentation, but I consider none of these as a requirement for publication:
- https://github.com/HuttleyLab/DiverseSeq/issues/135
- https://github.com/HuttleyLab/DiverseSeq/issues/136
- https://github.com/HuttleyLab/DiverseSeq/issues/137
- https://github.com/HuttleyLab/DiverseSeq/issues/138
- https://github.com/HuttleyLab/DiverseSeq/issues/139

I also opened this pull request fixing some minor typos:
- https://github.com/HuttleyLab/DiverseSeq/issues/140
