### 2025-03-21 #7765 @xin-huang
Hello @fboehm ,

I believe @GavinHuttley has addressed my concerns. I would suggest moving the figures into the main text rather than placing them at the end of the paper. However, I understand that the current sample template places the figures at the end.
