### 2025-03-22 c88b234
DOC: remove -hp option from docs, relates to #135
[NEW] support turning of progress using environment variable
    DVS_HIDE_PROGRESS

Thanks @iimog for the suggestion!
