### 2025-03-22 60d9616
DOC: advice to reproducing the docs, relates to #139
[NEW] doing the full reconstruction of the docs requires extra
    components. Describe how to get this.
