### 2025-03-22 HuttleyLab/DiverseSeq#139 @GavinHuttley
fixed by updating the docs to include special instructions for reproducing the docs examples by doing pip install diver-seq[doc]
