### 2025-03-23 9893d88
DEV: add an `extra` install option and modify docs
