+++ docs/index.md
@@ -15,10 +15,13 @@ pip install diverse-seq
 > **NOTE**
 > If you experience any errors during installation, we recommend using [uv pip](https://docs.astral.sh/uv/). This command provides much better error messages than the standard `pip` command. If you cannot resolve the installation problem, please [open an issue](https://github.com/HuttleyLab/DiverseSeq/issues).
 
-#### Install to replicate the documentation
+#### The `extra` install option
 
-To reproduce the examples presented in this documentation, you need extra components. You can install these as follows
+To get the additional components for working within Jupyter notebooks, or just to reproduce all the examples in the documentation, do the install as follows
 
 ```
-pip install "diverse-seq[doc]"
+pip install "diverse-seq[extra]"
 ```
+
+> **Note**
+> The first usage of `diverse-seq` is slow because both `cogent3` and `diverse-seq` are compiling some functions. This is a one-time cost.
