+++ pyproject.toml
@@ -53,24 +53,14 @@ test = [
     "pytest-cov",
     "pytest-xdist",
 ]
-doc  = ["click",
-        "cogent3[extra]",
-        "ipykernel",
-        "ipython",
-        "ipywidgets",
-        "jupyter_client",
-        "jupyterlab",
-        "jupytext",
-        "kaleido",
-        "mkdocs>=1.6.1",
-        "markdown-exec[ansi]>=1.10.0",
-        "mkdocs-jupyter>=0.25.1",
-        "markdown>=3.7",
-        "nbconvert>5.4",
-        "nbformat",
-        "numpydoc",
-        "plotly",
-        ]
+doc  = [
+    "diverse_seq[extra]",
+    "mkdocs>=1.6.1",
+    "markdown-exec[ansi]>=1.10.0",
+    "mkdocs-jupyter>=0.25.1",
+    "markdown>=3.7",
+    "numpydoc",
+]
 dev = [
     "cogapp",
     "docformatter",
@@ -78,6 +68,19 @@ dev = [
     "ruff==0.11.0",
     "diverse_seq[test]",
     "diverse_seq[doc]",
+]
+extra = [
+    "cogent3[extra]",
+    "ipykernel",
+    "ipython",
+    "ipywidgets",
+    "jupyter_client",
+    "jupyterlab",
+    "jupytext",
+    "kaleido",
+    "nbconvert>5.4",
+    "nbformat",
+    "plotly",
 ]
 
 [project.scripts]
