### 2025-03-23 f1bf78f
ENH: exit if provided data has too few sequences
[NEW] checking >= 5 sequences in the dvseqs file, if not displays message
    and exits with ret_code 1

[NEW] in dvs nmost/max checks the dvseqs file has > n/min_size
    sequences, if not displays message and exits with ret_code 1
