### 2025-03-23 8a260d3
ENH: exit if prep command would produce too few seqs
