### 2025-03-23 fa56736
DOC: better help string for prep -s option
