### 2025-03-23 3a0f122
DOC: describe using uv to install and use via uv
Related to #136, thanks for the suggestion @iimog!
