### 2025-03-26 2cf2620
MAINT: fix code that was broken in latest loky update
[CHANGED] the latest loky version raises pickling errors on numpy.isnan unless referenced
    explicitly from the module
