### 2025-03-26 f3e63d8
DOC: added missing notebook for generating synthetic known figure
[CHANGED] made colourn scheme consistent with othger figure, relates to
    point 9 of #134
