### 2025-03-26 e5e53b6
DEV: added missing statsmodels dependency for paper generation
