### 2025-03-28 HuttleyLab/DiverseSeq#134 @GavinHuttley
Hi @iimog,

Thank you very much for your thorough review of the manuscript. Below are our responses to your comments.

> 1. [grammar]

Accepted the suggestion

> 2. maximum entropy only for k=1

Correct! We have made this explicit in the text.

> 3. improve the text clarity

We have rewritten the text to be clearer.

> 4. Discrepancy between indicated sequence percentage and sample size

This is poor language on our part. We have the modified the text to the following

"We defined four sequence families that differ in nucleotide composition and two distinct "pools": *balanced*, in which there were 25 members from each sequence family, or *imbalanced*, where one sequence family occurred once, another 49 times and the remaining two families 25 times each. There were 50 replicates of each scenario."

> 5. figure 5 not referenced and no figure 1

We have fixed both of these.

> 6. be more explicit about what we mean by "relationship"

We have accepted the suggestion.

> 7. improve math notation

We have accepted the suggestion.

> 8. explain COV and stdev in the legend

We have accepted the suggestion.

> 9. make the solour scheme consistent

We have accepted the suggestion.
