### 2025-04-06 #7765 @GavinHuttley
@fboehm done
    - [x] latest version of `diverse-seq` is 2025.3.23
    - [x] https://doi.org/10.5281/zenodo.15164519
