+++ pyproject.toml
@@ -65,7 +65,7 @@ dev = [
     "cogapp",
     "docformatter",
     "flit",
-    "ruff==0.11.2",
+    "ruff==0.11.5",
     "diverse_seq[test]",
     "diverse_seq[doc]",
 ]
