### 2025-04-20 #7765 @fboehm
@GavinHuttley  - the pdf looks really good. A few minor suggestions are:

line 10 of pdf: "correspond" should be "corresponds"
line 43 of pdf: cite the "published approaches". Probably need to cite at least two here.
line 172 of pdf: does the mammals data set need a reference?
