### 2025-04-20 #7765 @fboehm
@GavinHuttley - it looks like the authors lists for 1) the pdf and 2) the archive differ. Please fix this. I see the archive has four authors: [GavinHuttley](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22GavinHuttley%22)

[Robert McArthur1](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Robert+McArthur%22)

[Katherine Caley1](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Katherine+Caley%22)

[Markus J. Ankenbrand2](https://zenodo.org/search?q=metadata.creators.person_or_org.name:%22Markus+J.+Ankenbrand%22)

while the pdf has just three right now.
