### 2025-04-22 #7765 @GavinHuttley
@fboehm thanks for your comments! Actions taken below.

## manuscript comments

> line 10 of pdf: "correspond" should be "corresponds"

Agreed and changed.

> line 43 of pdf: cite the "published approaches". Probably need to cite at least two here.

Agreed and changed.

> line 172 of pdf: does the mammals data set need a reference?

This data set was assembled for this work and is described on page 148. To make it clearer, we label this the "mammal data set" and refer back to this at the location you identified.

## artifact comments

>  it looks like the authors lists for 1) the pdf and 2) the archive differ

As per the comments of @iimog above, I have removed his name for the author list of the archive. Shame really, the contribution is appreciated!

> ensure that the archive title matches exactly that of the pdf.

The archive and title were automatically generated by Zenodo from the GitHub tag commit. [I have now changed the title](https://zenodo.org/records/15164520) as per the request.

> the version for the archive needs to match that of the repo. It looks like the archive has version "paper-accept", while the version that you listed above is "2025.3.23".

I added automatic Zenodo archive creation on producing new GitHub "releases" on realising we needed to have the archive for this paper. I triggered this specific archive by just making a "fake" release on GitHub (but not pushing to PyPI) by making an additional GitHub tag. The actual code is the same as is the code version string.

If this is critical, I can bump the version and do a new full release. I hesitate because that seems wasteful of resources.
