### 2025-04-25 #7765 @GavinHuttley
@fboehm - I have looked extensively for both of those and have not succeeded. The first (Biometry) is a book published by a company that no longer exists. McMillan took them over but searching for Biometry [on their site](https://www.macmillanlearning.com/college/us/search/?text=biometry) returns 0 hits. The second is for a chapter in a proceedings from the mid 1980's. While [I found an ISBN](https://searchworks.stanford.edu/view/3255202) for this issue attempting a DOI lookup via crossref was unsuccessful.
On this basis I conclude that they do not have DOIs.
