### 2025-04-26 #7765 @GavinHuttley
## Addressing issues with archive / version consistency

>  It looks like the archive's listing of your name has no space between first and last name. You should also add your affiliation to the archive. right now it looks like only the other two authors have affiliations.

[Done.](https://zenodo.org/records/15285722)

> ... we do need the version names (or numbers) to be the same for the repository and the archive before publishing the submission

New released version: 2025.3.24
Updated archive is: 10.5281/zenodo.15164519

@fboehm address final issues. (Apologies for the above attempts to use editorialbot to update those myself.)
