+++ .github/workflows/ci-cogent3-dev.yml
@@ -21,7 +21,7 @@ jobs:
             python-version: "3.12"
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v5
+        uses: astral-sh/setup-uv@v6
         with:
             enable-cache: true
 
