+++ .github/workflows/release.yml
@@ -20,7 +20,7 @@ jobs:
             python-version: "${{ matrix.python-version }}"
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v5
+        uses: astral-sh/setup-uv@v6
         with:
           enable-cache: true
           cache-dependency-glob: "pyproject.toml"
