### 2025-05-08 #7765 @fboehm
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.1371/journal.pone.0221068 is OK
> - 10.1074/mcp.T600022-MCP200 is OK
> - 10.1093/NAR/18.20.6097 is OK
> - 10.1109/18.61115 is OK
> - 10.1073/pnas.91.4.1455 is OK
> - 10.1093/nar/gkad1049 is OK
> - 10.1038/s41467-019-13443-4 is OK
> - 10.1186/gb-2007-8-8-r171 is OK
> - 10.1080/01621459.1979.10481038 is OK
> - 10.1093/oxfordjournals.molbev.a040454 is OK
> - 10.1038/ismej.2016.168 is OK
> - 10.1038/nbt.4229 is OK
> - 10.1145/2833157.2833162 is OK
> - 10.1093/molbev/msaa015 is OK
> - 10.1002/widm.53 is OK
> - 10.1186/s13059-016-0997-x is OK
> - 10.5281/zenodo.15098646 is OK
> - 10.5281/zenodo.15094004 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Biometry
> - No DOI given, and none found for title: Some probabilistic and statistical problems in the...
>
> ❌ MISSING DOIs
>
> - None
>
> ❌ INVALID DOIs
>
> - None
> ```

the two "skip dois" entries are, I believe articles that have no doi, presumably due to their ages. It's not a problem.
