### 2025-05-11 #7765 @GavinHuttley
Hi @@fboehm

Sorry for the hassle in getting the manuscript through the acceptance process!

I think the error was because of a markdown table formatting issue. I have committed a fix and checked the produced jat, confirming it does not contain an empty table. You can try again.

One question, should I place the tables / figures in the markdown file or will that be handled by the bot?

thanks again for your efforts!
