### 2025-06-05 7efc56a
MAINT: migrate to using cogent3 new_type core objects
