### 2025-06-06 b5d0e29
DEV: bug on windows for numba and python 3.13.4
[CHANGED] explicitly use 3.13.3 for GitHub actions for now
