+++ .github/workflows/ci.yml
@@ -24,6 +24,9 @@ jobs:
         with:
             python-version: "${{ matrix.python-version }}"
 
+      - name: Set UV_PYTHON environment variable
+        run: echo "UV_PYTHON=$(python -c 'import sys; print(sys.executable)')" >> $GITHUB_ENV
+
       - name: Install uv
         uses: astral-sh/setup-uv@v6
         with:
