+++ .github/workflows/ci.yml
@@ -24,9 +24,6 @@ jobs:
         with:
             python-version: "${{ matrix.python-version }}"
 
-      - name: Set UV_PYTHON environment variable
-        run: echo "UV_PYTHON=$(python -c 'import sys; print(sys.executable)')" >> $GITHUB_ENV
-
       - name: Install uv
         uses: astral-sh/setup-uv@v6
         with:
@@ -37,11 +34,12 @@ jobs:
       - name: "Installs for ${{ matrix.python-version }}"
         run: |
           uv venv venv -p python${{ matrix.python-version }}
-          uv tool install -p venv nox
+          uv pip install -p venv nox
 
       - name: "Run nox for ${{ matrix.python-version }}"
         shell: bash
         run: |
+          source venv/bin/activate
           cov="--cov-report lcov:lcov-${{matrix.os}}-${{matrix.python-version}}.lcov --cov-report term --cov-append --cov diverse_seq"
           nox -db uv --force-python ${{ matrix.python-version }} -s test -- $cov
 
