+++ .github/workflows/ci.yml
@@ -34,17 +34,13 @@ jobs:
       - name: "Installs for ${{ matrix.python-version }}"
         run: |
           uv venv venv -p python${{ matrix.python-version }}
-          uv pip install -p venv nox
+          uv tool install -p venv nox
 
       - name: "Run nox for ${{ matrix.python-version }}"
         shell: bash
         run: |
-          if [ "${{ runner.os }}" == "Windows" ]; then
-            source venv/Scripts/activate
-          else
-            source venv/bin/activate
-          fi
           cov="--cov-report lcov:lcov-${{matrix.os}}-${{matrix.python-version}}.lcov --cov-report term --cov-append --cov diverse_seq"
+          uv python pin ${{ matrix.python-version }}
           nox -db uv --force-python ${{ matrix.python-version }} -s test -- $cov
 
       - name: Coveralls Parallel
