### 2025-06-06 60ef99e
DEV: I give up
Excluding Windows/3.13 from being tested until this issue
is resolved by numba.
