### 2025-06-06 1a73a65
DEV: fix yaml syntax
