+++ .github/workflows/ci.yml
@@ -13,9 +13,9 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.10", "3.11", "3.12", "3.13"]
-      exclude:
-        - os: windows-latest
-          python-version: "3.13"
+        exclude:
+          - os: windows-latest
+            python-version: "3.13"
 
     steps:
       - uses: "actions/checkout@v4"
