### 2025-06-07 e3c4c87
DOC: add statement about using new_type
